% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{all_types}
\alias{all_types}
\title{Find links between formula and observation types}
\usage{
all_types(graph)
}
\arguments{
\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A matrix with formula types as rows and observation types as columns: "1" indicates
a link, "0" indicates no link.
}
\description{
Some observation types may belong to several different formula types, and some formula types
may give rise to several observation types.
}
