% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{arrange_data}
\alias{arrange_data}
\title{Arrange data set}
\usage{
arrange_data(data, graph)
}
\arguments{
\item{data}{A data frame with 3 columns: patient (numbering or names for each patient), time (the time when a patient was observed),
state (the state which the patient occupies at the observation time).}

\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A data frame with as many rows as there are patients and with one column per relevant time-point pluss one column indicating
the observation type of each patient.
}
\description{
Discard unrelevant time-points and arrange data into one row per patient format. The column names for the relevant time-points
are on the format t_{im}/t_{iM}, where i is the state numbering according to state_ordering() and m/M indicate the minimum and maximum
time-point in each state. For initial states only the maximum time-point is relevant, and for absorbing states only the minimum.
}
