% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_likelihood_func.R
\name{change_integrand}
\alias{change_integrand}
\title{Change integrand}
\usage{
change_integrand(integr)
}
\arguments{
\item{integrand}{An R function from eval-parse and type_to_integrand().}
}
\value{
An R function
}
\description{
Change integrand syntax from the form appropriate for repintegrate() to the form appropriate
for cubintegrate().
}
