% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{construct_formula_types}
\alias{construct_formula_types}
\title{Find all formula types}
\usage{
construct_formula_types(graph)
}
\arguments{
\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A vector with string elements indicating the states that were visited/traveled.
}
\description{
The formula types are the combination of states a patient can be last observed in and possible paths from
the initial states to these “last observed” states. The list of states is the path actually traveled by the patient,
but some of the transient state might actually be unobserved.
}
