% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{construct_obs_types}
\alias{construct_obs_types}
\title{Find all observation types}
\usage{
construct_obs_types(graph)
}
\arguments{
\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A vector with string elements indicating the states in which the patient is observed.
}
\description{
The observation types are the combination of states in which a patient can
actually observed. It does not necessarily indicate which path was traveled:
some observation types may correspond to multiple paths.
}
