% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{cubint}
\alias{cubint}
\title{Helper function for differentiating cubintegrate functions}
\usage{
cubint(integrand, lower, upper, tmax, param, xval)
}
\arguments{
\item{integrand}{The function to integrate.}

\item{lower}{Lower limits.}

\item{upper}{Upper limits.}

\item{tmax}{Time-points in which to evaluate the integrand.}

\item{param}{The parameter value in which to evaluate the integrand.}

\item{xval}{A vector of covariate values.}
}
\value{
The value after integration.
}
\description{
Helper function for differentiating cubintegrate functions
}
