% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{edge_matrices}
\alias{edge_matrices}
\title{Make edge matrices}
\usage{
edge_matrices(graph)
}
\arguments{
\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A list of three matrices: travelled, passedBy, and possible next.
}
\description{
Make three matrices with the formula types as rows, and edges in the graph as columns. For each formula type,
non-zero numbers indicate that edges have been "travelled", or "passed by", or if they are "possible next",
meaning that they may be travelled in the next step. The non-zero numbers also indicate in what order the
edges have been encountered. Zeros indicate edges that were not travelled/passed by/ possible next.
}
