% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_likelihood_func.R
\name{finding_limits}
\alias{finding_limits}
\title{Find the limits of integration}
\usage{
finding_limits(
  timepoints,
  form_type,
  edge_mats,
  absorbing_states,
  abs_exact = TRUE
)
}
\arguments{
\item{timepoints}{A named vector of relevant timepoints (belonging to a patient), i.e. a single
row from the output of the arrange_data() function.}

\item{form_type}{A formula type (as a string).}

\item{edge_mats}{A list with 3 matrices produced by the edge_matrices() function.}

\item{absorbing_states}{A vector denoting the absorbing states (in the ordered naming system).}

\item{abs_exact}{A boolean indicating whether the time of entrance into absorbing states is observed
exactly (TRUE) or not (FALSE). Default value is TRUE.}
}
\value{
A list with "lower" a vector of the lower limits of integration, "upper" a vector of the
upper limits of integration, and "tmax" the last timepoint (in which the integrand often needs
to be evaluated).
}
\description{
For a given formula type, a vector of (relevant) timepoints is sorted into the appropriate
integral limits.
}
