% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_functions}
\alias{get_functions}
\title{Get functions from the global environment}
\usage{
get_functions(target_env)
}
\arguments{
\item{target_env}{The environment to which the functions will be returned.}
}
\value{
All functions from the global environment.
}
\description{
Retrieves functions defined in the global environment and
returns them to the target environment.
}
\author{
Waldir Leoncio
}
