% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{hessian_matrix}
\alias{hessian_matrix}
\title{Calculates the hessian matrix (observed Fisher information matrix).}
\usage{
hessian_matrix(
  param,
  data,
  graph,
  X = NULL,
  mc_cores = 1,
  cmethod = "hcubature"
)
}
\arguments{
\item{param}{The parameter estimates from optimizer$par in smms.}

\item{data}{A data frame where the number of rows correspond to the total number of observations for all patients,
and  with 3 columns: patient (numbering or names for each patient), time (the time when a patient was observed),
state (the state which the patient occupies at the observation time). FOR NOW!!}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{X}{A matrix with the covariates. Number of rows equal to the number of patients and one column
for each covariate. The covariate specification is given by the user-specified densities.}

\item{mc_cores}{The number of cores to use (for parallelisation). The function uses the mclapply()
function from the parallel package. Defaults to 1.}

\item{cmethod}{The integration method of choice for the cubintegrate() function. Only for integrals
of higher dimension than 2. Defaults to "hcubature".}
}
\value{
A matrix with the hessian matrix.
}
\description{
Calculates the hessian matrix (observed Fisher information matrix).
}
