% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{occupancy_prob_delta}
\alias{occupancy_prob_delta}
\title{Calculating partial derivatives of occupancy probability functions wrt parameters}
\usage{
occupancy_prob_delta(state, time, param, graph, xval = NULL)
}
\arguments{
\item{state}{A string with a number indicating the state for which to compute the probability (in the ordered naming system).}

\item{time}{A vector of timepoints for which to compute the occupancy probability.}

\item{param}{The parameter values in which the probabilities should be evaluated. The dimension
and ordering is given by the user-specified densities.}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{xval}{A vector of covariate values.}
}
\value{
A vector of the same length as time.
}
\description{
Helper function for computing uncertainty bands.
}
