% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{overall_survival_ci_band}
\alias{overall_survival_ci_band}
\title{Compute uncertainty bands}
\usage{
overall_survival_ci_band(time, param, graph, xval, hessian, level = 0.95)
}
\description{
Compute uncertainty bands
}
