% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_preprocessing_func.R
\name{relevant_timepoints}
\alias{relevant_timepoints}
\title{Filter away redundant observations in a multi-state dataset}
\usage{
relevant_timepoints(data, graph)
}
\arguments{
\item{data}{A data frame with 3 columns: patient (numbering or names for each patient), time (the time when a patient was observed),
state (the state which the patient occupies at the observation time).}

\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}
}
\value{
A data frame with the same 3 columns as 'data', but with fewer rows (potentially).
}
\description{
Only the last timepoint in initial states are relevant. Only the first timepoint in absorbing states. For the
transient states the first and last timepoints are relevant.
}
