% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_likelihood_func.R
\name{repint2}
\alias{repint2}
\title{Integrate functions of dimension 1 and 2}
\usage{
repint2(ss, innerfunc, tt, param, lower2, upper2, x, tt2 = -1, ...)
}
\arguments{
\item{innerfunc}{The integrand (an R function), with inputs tt, ss, and uu (if dimension=2).}
}
\value{
The value of the integral.
}
\description{
The following two functions are helper functions for integrating in dimension 1 or 2.
}
