% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_likelihood_func.R
\name{repintegrate}
\alias{repintegrate}
\title{Integrate functions of dimension 1}
\usage{
repintegrate(innerfunc, tt, param, x, lower, upper, tt2 = -1, ...)
}
\arguments{
\item{innerfunc}{The integrand (an R function).}

\item{tt}{An input to innerfunc (a time-point in which to evaluate the integrand).}

\item{param}{The parameter value in which to evaluate the integrand.}

\item{x}{A vector of covariate values.}

\item{lower}{Lower limits.}

\item{upper}{Upper limits.}

\item{tt2}{An input to innerfunc (a time-point in which to evaluate the integrand).}

\item{...}{further arguments to innerfunc.}
}
\value{
The value of the integral.
}
\description{
The following is a helper function for integrating in dimension 1.
}
