% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{transition_prob_ci_band}
\alias{transition_prob_ci_band}
\title{Compute the transition probability over time with uncertainty bands}
\usage{
transition_prob_ci_band(
  trans_ji,
  time_t,
  time_v,
  param,
  graph,
  xval,
  hessian,
  level = 0.95
)
}
\arguments{
\item{trans_ji}{A string with two state names separated by "-" indicating the prior state j and the current state i (using user-defined state names).}

\item{time_t}{A vector of timepoints denoting the "current" time.}

\item{time_v}{A number indicating the single prior time-point (must be smaller than all time_t).}

\item{param}{The parameter values in which the probabilities should be evaluated. The dimension
and ordering is given by the user-specified densities.}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{xval}{A vector of covariate values.}

\item{hessian}{The hessian matrix from the optimisation of the negative log-likelihood.}

\item{level}{The confidence level for the uncertainty bands, defaults to 0.95.}
}
\value{
A vector of the same length as time.
}
\description{
Compute the transition probability over time with uncertainty bands
}
