% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{transition_prob_delta}
\alias{transition_prob_delta}
\title{Calculating partial derivatives of transition probability functions wrt parameters}
\usage{
transition_prob_delta(trans_ji, time_t, time_v, param, graph, xval = NULL)
}
\arguments{
\item{trans_ji}{A string with two state names separated by "-" indicating the prior state j and the current state i (using user-defined state names).}

\item{time_t}{A vector of timepoints denoting the "current" time.}

\item{time_v}{A number indicating the single prior time-point (must be smaller than all time_t).}

\item{param}{The parameter values in which the probabilities should be evaluated. The dimension
and ordering is given by the user-specified densities.}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{xval}{A vector of covariate values.}
}
\value{
A vector of the same length as time.
}
\description{
Helper function for computing uncertainty bands.
}
