% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_likelihood_func.R
\name{type_to_integrand}
\alias{type_to_integrand}
\title{Write out the integrand as a string}
\usage{
type_to_integrand(form_type, edge_mats, names_surv_dens, abs_exact = TRUE)
}
\arguments{
\item{form_type}{A formula type (as a string).}

\item{edge_mats}{A list with 3 matrices produced by the edge_matrices() function.}

\item{names_surv_dens}{A data frame indicating the names of survival functions and densities.
The output of the names_of_survival_density() function.}

\item{abs_exact}{A boolean indicating whether the time of entrance into absorbing states is observed
exactly (TRUE) or not (FALSE). Default value is TRUE.}
}
\value{
A text string giving the R syntax of the integrand.
}
\description{
For a given formula type the appropriate integrand is generated as a text string. The integrand
takes the form of an R type function. For now this function is limited to integrals up to dimension 5.
}
