% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{variance_matrix}
\alias{variance_matrix}
\title{Calculates the variance-covariance matrix.}
\usage{
variance_matrix(param, data, graph, X = NULL, mc_cores = 3)
}
\arguments{
\item{param}{The parameter estimates from optimizer$par in smms.}

\item{data}{A data frame where the number of rows correspond to the total number of observations for all patients,
and  with 3 columns: patient (numbering or names for each patient), time (the time when a patient was observed),
state (the state which the patient occupies at the observation time). FOR NOW!!}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{X}{A matrix with the covariates. Number of rows equal to the number of patients and one column
for each covariate. The covariate specification is given by the user-specified densities.}

\item{mc_cores}{The number of cores to use (for parallelisation). The function uses the mclapply()
function from the parallel package.}
}
\value{
A matrix with the variance-covariance matrix.
}
\description{
Calculates the variance-covariance matrix.
}
