% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_latex_func.R
\name{write_loglikelihood}
\alias{write_loglikelihood}
\title{Write log-likelihood for a graph in latex format.}
\usage{
write_loglikelihood(graph, abs_exact = TRUE)
}
\arguments{
\item{graph}{A directed, acyclic graph in the igraph format (igraph package).}

\item{abs_exact}{A boolean indicating whether the time of entrance into absorbing states is observed
exactly (TRUE) or not (FALSE). Default value is TRUE.}
}
\value{
Writes a txt file into the working directory within which one will find the entire log-likelihood
formula belonging to the graph, in latex format.
}
\description{
Write log-likelihood for a graph in latex format.
}
