% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{est_ci}
\alias{est_ci}
\title{Estimates and confidence intervals}
\usage{
est_ci(param, hessian, level = 0.95, log = TRUE)
}
\arguments{
\item{param}{The parameter estimates.}

\item{hessian}{The hessian matrix.}

\item{level}{The level of confidence. Default value is 0.95.}

\item{log}{Boolean denoting whether the intervals should be on the log scale (default), or on the exp/hazard scale.}
}
\value{
A vector of the same length as time.
}
\description{
Provide estimates and approximate confidence intervals for all parameters
(on the "original" scale - meaning that all parameters live on -Inf to +Inf).
}
