% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{occupancy_prob_ci_band}
\alias{occupancy_prob_ci_band}
\title{Compute uncertainty bands}
\usage{
occupancy_prob_ci_band(state, time, param, graph, xval, hessian, level = 0.95)
}
\description{
Compute uncertainty bands
}
