% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{smms}
\alias{smms}
\title{Fit a semi-Markovian multistate model}
\usage{
smms(startval, data, graph, X = NULL, mc_cores = 3, variance_matrix = FALSE)
}
\arguments{
\item{startval}{The parameter values where the optimisation routine will start. The dimension
and ordering is given by the user-specified densities.}

\item{data}{A data frame where the number of rows correspond to the total number of observations for all patients,
and  with 3 columns: patient (numbering or names for each patient), time (the time when a patient was observed),
state (the state which the patient occupies at the observation time). FOR NOW!!}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{X}{A matrix with the covariates. Number of rows equal to the number of patients and one column
for each covariate. The covariate specification is given by the user-specified densities.}

\item{mc_cores}{The number of cores to use (for parallelisation). The function uses the mclapply()
function from the parallel package.}

\item{variance_matrix}{Whether the variance-covariance matrix for the parameter estimates should be calculated
or not.}
}
\value{
The result from optimising the log-likelihood: parameter estimates with corresponding variance-covariance
matrix if variance_matrix = TRUE, and the maximum log-likelihood value.
}
\description{
The function the user should interact with. Assumes that the user has provided densities and
survival functions of the right format (ideally with all parameters living on -Inf to Inf -
to avoid having to set lower limits for some parameters).
}
