% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_msm_optim_func.R
\name{transition_prob}
\alias{transition_prob}
\title{Compute the transition probability over time}
\usage{
transition_prob(trans_ji, time_t, time_v, param, graph, xval = NULL)
}
\arguments{
\item{trans_ji}{A string with two state names separated by "-" indicating the prior state j and the current state i (using user-defined state names).}

\item{time_t}{A vector of timepoints denoting the "current" time.}

\item{time_v}{A number indicating the single prior time-point (must be smaller than all time_t).}

\item{param}{The parameter values in which the probabilities should be evaluated. The dimension
and ordering is given by the user-specified densities.}

\item{graph}{A directed, acyclic graph giving the multistate structure, in the igraph format (igraph package).}

\item{xval}{A vector of covariate values.}
}
\value{
A vector of the same length as time_t.
}
\description{
The probability that a patient is found in state i at a time-point t given that she was in state j at
a prior time-point v. Returns a vector of zeros if the transition from j to i is not possible (given the graph).
}
\details{
Only computes probabilities for direct transitions (a specific edge in the graph)
}
